Split tests into separate files for minio and http
